refactor: group card image provider
