+++ apps/backend/src/chat/cardImages/index.ts
@@ -2,6 +2,6 @@ export { generateCardImage } from "./operation";
 export {
   generatedCardImageModel, generatedCardImageOutputFormat,
   generatedCardImageQuality, generatedCardImageSize,
-} from "./openaiAdapter";
+} from "./provider/openaiAdapter";
 export type { GeneratedCardImageInput, GeneratedCardImageResult } from "./types";
 export type { GeneratedCardImageObservationContext } from "./providerTypes";
